fix: export proxyRefs
close #12600
